requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
flake8==6.0.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.1
psycopg2==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==1.7.1
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
